Fix coop demos sometimes crashing the game on playback
This is a fun one.

SAR injects information into demos in a pretty janky way. In order to
maintain compatibility with the base game, we hijack the
RadialMouseMenuCallback CustomData message in demos, sending a longer
payload containing SAR data past the expected 8 bytes. This isn't
pretty, but it works fine so long as the length assertions for the
callback aren't compiled in, and it maintains compatibility with the
base game quite nicely.

However, CustomData packets work a bit weirdly. At the start of the
demo, a special packet - with custom data type -1 - is written, which
contains records of every other type (since they're given by numeric ID
in the demo, but handled by name in the engine). I always assumed that
this packet was written as soon as the demo started. However, it turns
out that isn't necessarily true; instead, this packet (and most of the
stuff that should be at the start of a demo) is written when the
signonstate switches to FULL, whereas demo recording starts a bit
earlier, on SIGNONSTATE_NEW.

Most of the time, this would be a nonissue; we have no reason to write
any information to demos in this period. However, as usual, janky coop
comes along and makes everything worse. At the beginning of a coop map,
while Orange is loading in, Blue is in a strange state; they're
effectively fully in-game - they can look around etc - but their
signonstate isn't yet FULL, so the demo information hasn't been written.
Crucially, the server - and hence entity simulation - is still happening
at this time. This causes an edge case; because entity inputs can happen
(since the server is active), it is possible for us to try and record an
entity input to the demo.

At that point, we write a CustomData message to the demo *before* the
message that tells it how to actually handle said CustomData. The game
has no checks here (at least not in the prod build), so it theoretically
just crashes; however, the handler information from previously played
demos is actually retained in a global, so playing any other demo (they
all have the same handler info in Portal 2) mitigates the issue. But the
demo itself is still unambiguously broken; we're only able to play it
through a quirk of the demo player.

The fix here is quite simple; don't write any custom data to demos
between SIGNONSTATE_NEW and SIGNONSTATE_FULL.