Improve text rendering on Linux
CMatSystemSurface::DrawTextLen is pretty borked on Linux. It doesn't
seem to account for variable kerning properly; it assumes a character
takes up a constant width on the screen, which only works for a few font
indices. This replacement isn't perfect, but it seems to work faiiirly
well - it only ever overestimates, and that seems quite rare.