Fix crash in AcceptInput hook on Windows
Because the AcceptInput method is called according to the thiscall
convention whereas its hook is called according to cdecl, we have to
copy thisptr from ecx onto the stack. cdecl functions may clobber ecx,
so we have to pop it back off later to restore it for the main function
body. Previously, this was done by simply popping the first argument
back into ecx. However, cdecl functions are allowed to clobber their
stack arguments; as such, the optimised Windows build was crashing when
this hook was run in seemingly-random cases, as changes to AcceptInput
were causing it to clobber this stack argument. Instead, we now push ecx
*first*, then all the args (including ecx again), then pop all the args
and re-pop ecx.