[renderer] prevent double mutex unlock race condition
The hook for SND_RecordBuffer could, when a render was finished (i.e.
isRendering became false), attempt to unlock a mutex which it had
already unlocked. Because we finish rendering asynchronously (on the
worker thread) unless we absolutely need to synchronise (e.g. to start
another render), this could happen whenever a render was finished. This
is UB per spec, and was likely the cause of some random crashes on
Windows.