fix: don't use net time in demo playback
There's actually a bigger underlying problem here, but I don't care
enough to fix it atm. Basically, if the game is timescaled in some way
(e.g. host_framerate or rendering), net_time still runs based off of
real time, which causes the session tick to be completely inaccurate.
I'm pretty sure this problem can manifest itself everywhere, but it's
going to be most of an issue in demo playback (e.g. when rendering), so
for now I've just switched to always using the engine tick count in demo
playback.