refactor: fix hook signatures on Linux
This weirdness was caused by section 3.1.2.3 of the Itanium C++ ABI
(followed by GCC). It states that any parameter considered "non-trivial
for the purposes of calls" is passed by reference, and that any
parameter with a non-trivial copy constructor is considered as such.
Because we don't replicate the complexity of actual datastructures, we
missed out a copy constructor, making a couple of hooks seem wrong on
Linux.