feat: `sar_frametime_uncap`
EXPERIMENTAL - USE AT OWN RISK
This cvar is likely to break a number of things when enabled. It is not yet confirmed to be allowed at all!

moved it behind a cvar because it has a lot of unforeseen side-effects. also added debug stuff! should it be part of `sar_tick_debug`?
this hook could be used to improve `sar_hud_fps` but who uses that anyway