refactor: writing filepaths
if a file already exists somewhere, overwrite it
i'm not a huge fan of this but it's still better
than putting it in arbitrary folders. if you, for
example, dump cvars then move it to your p2common,
subsequent dumps will overwrite the file in
p2common instead of creating a new one in the
game directory.

this also touches TAS code (scary) so i hope it
works properly (i haven't tested the protocol
script thing)