fix: twtm, aptag, and linux
The .so uses some .bs (haha) to get to the thing,
and I don't fully understand it yet. For now,
Portal 2 is supported, and the mods will just not
have the feature.

Also speedrun mod seems to not have the feature
either, after adding safety checks, so I will see
about that tomorrow.