feat: support for 2011 Portal 2
Minimum effort to make things work on the April 27th, 2011 build of the
game. LP hud is broken (I'm guessing the field for the Portal count was
added with the CM update?). TAS tools won't work due to
SteamControllerMove not existing at all, there's also a crash around
line 384 or so of TasPlayer.cpp regardless when trying to play a TAS
that I didn't care enough to figure out. SAR's strategy for injecting
data into demos won't work either, as this version of the game doesn't
save the radial menu cursor position to the demo.

Otherwise, I was able to play a full AMC run with bill, record/play
demos, render a demo, and do other things without issues.

I cannot totally verify the integrity of the game is intact (when it
comes to potentially incorrect vtable offsets or the like causing wrong
functions to be hooked), but I will leave that to someone more familiar
with the codebase/features to look into. Anyways, this was a fun way to
spend like 5 entire afternoons/evenings.