fixup: clean up some trolling from the previous PR
This was meant to be a force push on the previous PR, but then the PR
was merged 8 minutes before I went to go message AMJ to tell him not to
merge yet (oops lol).

Anyways, made the thing in LPHud early return, demystified some memes
with the MsgFunc_SayText(2) hook, and mostly undid/cleaned up the
changes to Interface::GetPtr in favor of a simpler solution. (Related to
that, the console module is now initialized first before any other
modules, so that it is available immediately within the plugin.)