build: migrate to `zig build`
This commit gets the Linux build of SAR working via `zig build` on Zig
0.14.0 (and current dev builds). This solves the problems we've been
having with glibc versions, because Zig is able to provide far older
glibc versions than many toolchains upon request. I am somewhat
arbitrarily targeting glibc 2.28, which seems to work.

In theory, Zig *could* be used to build SAR for Windows, although
it's more finickey than on Linux because of ABI differences between the
`gnu` and `msvc` ABIs. SAR relies on the MSVC C++ ABI on Windows in
order for vtable and data layout to match the game modules, so we need
to target `x86-windows-msvc`. This unfortunately makes cross-compilation
essentially impossible, because Zig cannot distribute the MSVC library
files.

As a part of this commit, several libraries are now built on-the-fly
rather than being linked as binary blobs in the repo. These libraries
are:

* libx265 (a dependency of ffmpeg, for sar_render)
* SFML
* discord-rpc

The reason for this is that these three libraries contain C++ code, and
using Zig's build system means we are using LLVM's libc++ instead of
GCC's libstdc++. These c++ stdlib implementations have different ABIs,
so any code linking to libc++ -- i.e. any C++ code -- needed
recompiling. Given that I was going to have to rebuild these
dependencies anyway, I figured it made sense to actually integrate their
builds.

These dependencies are not vendored. Instead, Zig's package manager is
used to fetch tarballs on-the-fly. Mostly, these tarballs come directly
from upstreams; SFML is an exception, because their 2.5.x branch needed
changes to build with modern libc++, so the URL in `build.zig.zon` is a
fork with minor tweaks.

Long-term, it would make sense to migrate *all* of our dependencies to
work like this, and on Windows too; vendoring binary blobs in the repo
is pretty clearly a Bad Thing. However, that's out of scope for this PR.